+++ barney/DeviceGroup.h
@@ -92,7 +92,7 @@ namespace barney {
              int globalIndexStep);
     ~DevGroup();
     
-    int size() const { return devices.size(); }
+    int size() const { return (int)devices.size(); }
     
     // OWLGeomType getOrCreateGeomTypeFor
     // (const std::string &geomTypeString,
